+++ src/ir_JVC.hpp
@@ -59,7 +59,6 @@
 // LSB first, 1 start bit + 8 bit address + 8 bit command + 1 stop bit.
 // The JVC protocol repeats by skipping the header mark and space -> this leads to a poor repeat detection for JVC protocol.
 // Some JVC devices require to send 3 repeats.
-
 #define JVC_ADDRESS_BITS      8 // 8 bit address
 #define JVC_COMMAND_BITS      8 // Command
 
@@ -119,7 +118,7 @@ bool IRrecv::decodeJVC() {
 
     // Check we have the right amount of data (36 or 34). The +4 is for initial gap, start bit mark and space + stop bit mark.
     // +4 is for first frame, +2 is for repeats
-    if (decodedIRData.rawlen != ((2 * JVC_BITS) + 2) && decodedIRData.rawlen != ((2 * JVC_BITS) + 4)) {
+    if (!(decodedIRData.rawlen == ((2 * JVC_BITS) + 2) || decodedIRData.rawlen == ((2 * JVC_BITS) + 4))) {
         DEBUG_PRINT(F("JVC: Data length="));
         DEBUG_PRINT(decodedIRData.rawlen);
         DEBUG_PRINTLN(F(" is not 34 or 36"));
@@ -128,10 +127,10 @@ bool IRrecv::decodeJVC() {
 
     if (decodedIRData.rawlen == ((2 * JVC_BITS) + 2)) {
         /*
-         * Check for repeat
-         * Check leading space and first and last mark length
+         * 34 -> check for repeat distance and check leading space and first and last mark length
          */
-        if (decodedIRData.initialGapTicks < ((JVC_REPEAT_DISTANCE + (JVC_REPEAT_DISTANCE / 4) / MICROS_PER_TICK))
+        if (lastDecodedProtocol == JVC
+                && decodedIRData.initialGapTicks < ((JVC_REPEAT_DISTANCE + (JVC_REPEAT_DISTANCE / 4) / MICROS_PER_TICK))
                 && matchMark(irparams.rawbuf[1], JVC_BIT_MARK)
                 && matchMark(irparams.rawbuf[decodedIRData.rawlen - 1], JVC_BIT_MARK)) {
             /*
@@ -144,6 +143,7 @@ bool IRrecv::decodeJVC() {
         }
     } else {
 
+        // 36 -> check for new start frame
         if (!checkHeader_P(&JVCProtocolConstants)) {
             return false;
         }